Refactor reader API
